backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0